Add intermediate school meal variables
