+++ openfisca_us/variables/usda/school_meals/free_school_meals.py
@@ -12,4 +12,4 @@ class free_school_meals(Variable):
     def formula(spm_unit, period, parameters):
         tier = spm_unit("school_meal_tier", period).decode_to_str()
         is_free = tier == "FREE"
-        return is_free * spm_unit("school_meal_subsidy", period)
+        return is_free * spm_unit("school_meal_net_subsidy", period)
