+++ openfisca_us/variables/usda/school_meals/reduced_price_school_meals.py
@@ -12,4 +12,4 @@ class reduced_price_school_meals(Variable):
     def formula(spm_unit, period, parameters):
         tier = spm_unit("school_meal_tier", period).decode_to_str()
         is_reduced_price = tier == "REDUCED"
-        return is_reduced_price * spm_unit("school_meal_subsidy", period)
+        return is_reduced_price * spm_unit("school_meal_net_subsidy", period)
